Updates in main.py READ COMMENTS
Creation of a column 'word_lower'. Now the case errors will not be followed by a 'Word not found error'!
Remove some useless code
The main.py file should soon be removed when the gui.py will be operational. This file serves to test basic code in a command line form